requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
requests==2.31.0
pdfplumber==0.10.3
sentence-transformers==2.2.2
faiss-cpu==1.7.4
aiohttp==3.9.1
python-multipart==0.0.6
pydantic==2.5.0
numpy==1.24.3
